[Feature #160941788] Writes Tests for Endpoints
