+++ tests/v2/test_menu.py
@@ -3,24 +3,10 @@ import unittest
 import json
 from app import create_app, connect
 from .test_auth import create_admin_token
+from tests.base import BaseTestCase
 
-class TestDevelopmentConfig(unittest.TestCase):
-    def setUp(self):
-        self.app = create_app(config_name="testing")
-        self.app_context = self.app.app_context()
-        self.app_context.push()
-        self.client = self.app.test_client()
+class TestDevelopmentConfig(BaseTestCase):
 
-        self.correct_menu = {
-            "name": "Burger Tuple",
-            "price": 800,
-            "decription": "This is a burger Tuples"
-            }
-      
-        self.admin_token = create_admin_token()
-
-    def tearDown(self):
-        self.api_test_client = None
     
     def test_create_menu(self):
         """
